cleanup x3 and fixed trash behavior
