+++ scripts/index.js
@@ -8,7 +8,8 @@ import * as compute from './compute.js';
 import * as graph_ops from './graph_ops.js';
 import * as menus from './menus.js';
 import * as permalink from './permalink.js';
-// import * as home from './home.js';
+import * as util from './util.js';
+import * as ui_setup from './ui_setup.js';
 
 // if not in browser, don't run
 if (typeof document !== 'undefined') {
@@ -16,7 +17,7 @@ if (typeof document !== 'undefined') {
   window.addEventListener('resize', () => drawing.draw(graph));
 }
 
-let graph = consts.EMPTY_GRAPH;  // global graph
+let graph = {};  // global graph
 
 /** handles double click */
 function bind_double_click() {
@@ -163,11 +164,9 @@ function bind_drag() {
       if (clicked_edge) {   // left drag edge
         drag_edge = higher_order_drag_edge(clicked_edge);        
         canvas.addEventListener('mousemove', drag_edge);
-        trash_color();      // change color if edge is over trash
       } else if (clicked_vertex) {  // vertex has lower priority than edge
         drag_vertex = higher_order_drag_vertex(clicked_vertex);  // create the function
         canvas.addEventListener('mousemove', drag_vertex);
-        trash_color();      // change color if vertex is over trash
       } else {  // left drag scene
         canvas.addEventListener('mousemove', drag_scene);
       } 
@@ -179,30 +178,17 @@ function bind_drag() {
     const drop_edge = drawing.in_edge_text(graph, x, y);
     if (drop_vertex && drawing.over_trash(e)) { // delete vertex if dropped over trash
       graph_ops.delete_vertex(graph, drop_vertex);
-    }
-    else if (drop_edge && drawing.over_trash(e)) { // delete edge if dropped over trash
+    } else if (drop_edge && drawing.over_trash(e)) { // delete edge if dropped over trash
       graph_ops.delete_edge(graph, drop_edge);
+    } else if (drawing.over_trash(e)) {
+      delete_graph();  // delete the entire graph if dropped over trash
     }
-    else { // not dropped over trash can and release drag
-      canvas.removeEventListener('mousemove', drag_scene);
-      canvas.removeEventListener('mousemove', drag_vertex);
-      canvas.removeEventListener('mousemove', drag_edge);
-      canvas.removeEventListener('mousemove', edge_animation);
-      mutex = false;  // release the resource
-    }
-  });
-}
-
-/* adds mouse event to change the color of the trash if over the icon */
-function trash_color() {
-  const canvas = drawing.get_canvas();
-  canvas.addEventListener('mousemove', e => {
-    if (drawing.over_trash(e)) {
-      drawing.recolor_trash(true);
-    }
-    else {
-      drawing.recolor_trash(false);
-    }
+    
+    canvas.removeEventListener('mousemove', drag_scene);
+    canvas.removeEventListener('mousemove', drag_vertex);
+    canvas.removeEventListener('mousemove', drag_edge);
+    canvas.removeEventListener('mousemove', edge_animation);
+    mutex = false;  // release the resource
   });
 }
 
@@ -230,76 +216,6 @@ function bind_context_menu() {
       menus.display_edge_menu(graph, edge, e.clientX, e.clientY);
     }
   });
-}
-
-/** binds the run button to the run function */
-function bind_plus_minus() {
-  const plus_button = document.getElementById('plus_button');
-  const minus_button = document.getElementById('minus_button');
-
-  plus_button.addEventListener('click', () => {
-    const machine_inputs = document.querySelector('.machine_inputs');
-    //first, check if there is a hidden machine input
-    for (let i = 0; i < machine_inputs.children.length; i++) {
-      if (machine_inputs.children.item(i).classList.contains('machine_input') && machine_inputs.children.item(i).hidden) {
-        machine_inputs.children.item(i).hidden = false;
-        return;
-      }
-    }
-    // create a new machine input
-    add_input_bar();
-    bind_run_input();
-  });
-
-  // event listener for minus button
-  minus_button.addEventListener('click', () => {
-    // remove the last machine input
-    const machine_inputs = document.querySelector('.machine_inputs');
-    //iterate backwards through machine inputs, hide the first one that is not hidden
-    for (let i = machine_inputs.children.length - 1; i >= 0; i--) {
-      if (machine_inputs.children.item(i).classList.contains('machine_input') && !machine_inputs.children.item(i).hidden) {
-        machine_inputs.children.item(i).hidden = true;
-        break;
-      }
-    }
-  });
-}
-
-/** Generates a new input bar using the DOM
- * I tried using HTML to generate the first one
- * but ran into inconsistent spacing issue I couldn't figure out,
- * so this will be used to generate all for consistency's sake
- */
-function add_input_bar(){
-  const machine_inputs = document.querySelector('.machine_inputs');
-  const new_machine_input = document.createElement('div');
-  new_machine_input.classList.add('machine_input');
-
-  // create a new textarea box
-  const new_textarea = document.createElement('textarea');
-  new_textarea.classList.add("machineInput");
-  // create a new run button under the original one
-  const new_run_button = document.createElement('button');
-  new_run_button.classList.add('run_btn');
-  new_run_button.innerHTML = 'R';
-
-  // create a new step button
-  const new_step_button = document.createElement('button');
-  new_step_button.classList.add('step_btn');
-  new_step_button.innerHTML = 'S';
-
-  // create a new reset button
-  const new_reset_button = document.createElement('button');
-  new_reset_button.classList.add('reset_btn');
-  new_reset_button.innerHTML = 'X';
-
-  new_machine_input.appendChild(new_textarea);
-  new_machine_input.appendChild(new_run_button);
-  new_machine_input.appendChild(new_step_button);
-  new_machine_input.appendChild(new_reset_button);
-
-  // append the new button to the body
-  machine_inputs.appendChild(new_machine_input);
 }
 
 /** binds each machine input to the run_input function */
@@ -375,35 +291,18 @@ function bind_scroll() {
   });
 }
 
-/**
- * helper function to abstract away double clicking
- * @param {string} key - ex. KeyZ, KeyA
- * @param {Function} callback - a function to be called when double click happens
- */
-function on_double_press(key, callback) {
-  let last_time = 0;
-  document.addEventListener('keypress', e => {
-    if (e.code === key) {
-      if (e.timeStamp-last_time < consts.DOUBLE_PRESS_TIME) {
-        callback();
-        last_time = 0;  // prevent triple click
-      } else {
-        last_time = e.timeStamp;
-      }
-    }
-  });
+function delete_graph() {
+  if (!Object.keys(graph).length) {  // nothing to delete
+    return;
+  }
+  graph = {};
+  drawing.draw(graph);
+  hist.push_history(graph);
 }
 
 /** press dd does delete */
 function bind_dd() {
-  on_double_press('KeyD', () => {
-    if (!Object.keys(graph).length) {  // nothing to delete
-      return;
-    }
-    graph = consts.EMPTY_GRAPH;
-    drawing.draw(graph);
-    hist.push_history(graph);
-  });
+  util.on_double_press('KeyD', delete_graph);
 }
 
 function hash_change_handler() {
@@ -467,43 +366,6 @@ function bind_save_drawing() {
   save_btn.addEventListener('click', () => drawing.save_as_png(graph));
 }
 
-/** dynamically change the length of textboxes */
-export function bind_elongate_textbox() {
-  const change_width_func = e => {  // minimum width of 4ch
-    e.target.style.width = `${Math.max(4, e.target.value.length)}ch`;
-  };
-  document.querySelectorAll('input[type=text]').forEach(textbox => {
-    textbox.addEventListener('input', change_width_func);
-  });
-
-  // /** dynamically change the height of textbox */
-  // const resizeTextArea = textarea => {
-  //   const { style, value } = textarea;
-
-  //   // The 4 corresponds to the 2 2px borders (top and bottom):
-  //   // style.height = style.minHeight = 'auto';
-  //   // style.minHeight = `${ Math.min(textarea.scrollHeight + 4, parseInt(textarea.style.maxHeight)) }px`;
-  //   // style.height = `${ textarea.scrollHeight + 4 }px`;
-  //   // style.maxHeight = `100px`
-
-  // }
-
-  // const textarea = document.getElementById('machineInput');
-
-  // textarea.addEventListener('input', () => {
-  //   resizeTextArea(textarea);
-  // }); /* dynamically change the height of textbox */
-}
-
-
-
-
-// const textarea = document.getElementById('machineInput');
-
-// textarea.addEventListener('input', () => {
-//   resizeTextArea(textarea);
-// });
-
 /** button to generate permanent link */
 function bind_permalink() {
   const permalink_btn = document.getElementById('permalink');
@@ -523,239 +385,15 @@ function init() {
   bind_double_click();
   bind_drag();
   bind_context_menu();
-  bind_plus_minus();
-  add_input_bar(); // called so one input bar appears on opening of homepage
   bind_run_input();
   bind_machine_transform();
   bind_save_drawing();
   bind_undo_redo();
   bind_scroll();
   bind_dd();
-  bind_elongate_textbox();
   bind_permalink();
-  trash_color(); // link mouse event to appropriate trash color
-  htmlSetUp(); // initiate eventlisteners for sidenavbar, second sidenavbar, and popup tutorial
+  ui_setup.bind_plus_minus();
+  ui_setup.add_input_bar(); // called so one input bar appears on opening of homepage
+  ui_setup.htmlSetUp(); // initiate eventlisteners for sidenavbar, second sidenavbar, and popup tutorial
   init_graph();  // leave this last since we want it to override some of the above
-}
-
-/** moved basic set up from index.html and combined into one function */
-function htmlSetUp(){
-  // first time pop up implementation
-  // Check if the user is a first-time visitor
-  if (!localStorage.getItem('visitedBefore')) {
-  // User is a first-time visitor
-  openPopup();
-
-  // Set flag to indicate the user has visited before
-  localStorage.setItem('visitedBefore', true);
-  } 
-
-  const closeButton = document.getElementById('closeButton');
-  closeButton.addEventListener('click', () => {
-    closeMenu();
-  })
-
-  const homeIcon = document.getElementById('homeIcon');
-  const machineIcon = document.getElementById('machineIcon');
-  const saveIcon = document.getElementById('saveIcon');
-  const bugIcon = document.getElementById('bugIcon');
-  const helpIcon = document.getElementById('helpIcon');
-  const tutorial_close_btn = document.getElementById('tutorial_close_btn');
-  const tutorial_finish_btn = document.getElementById('tutorial_finish_btn');
-
-  homeIcon.addEventListener("click", () => {expandIcon('home')});
-  machineIcon.addEventListener("click", () => {expandIcon('settings')});
-  saveIcon.addEventListener("click", () => {expandIcon('save')});
-  bugIcon.addEventListener("click", () => {redirectToBugReport()});
-  helpIcon.addEventListener("click", () => {
-    openPopup();
-    closeMenu();
-  }); 
-  tutorial_close_btn.addEventListener("click",() => {closePopup();})
-  tutorial_finish_btn.addEventListener("click",() => {closePopup();})
-
-  const nextBtn_1to2 = document.getElementById('nextBtn_1to2');
-  const nextBtn_2to3 = document.getElementById('nextBtn_2to3');
-  const nextBtn_3to4 = document.getElementById('nextBtn_3to4');
-  const nextBtn_4to5 = document.getElementById('nextBtn_4to5');
-
-  const prevBtn_2to1 = document.getElementById('prevBtn_2to1');
-  const prevBtn_3to2 = document.getElementById('prevBtn_3to2');
-  const prevBtn_4to3 = document.getElementById('prevBtn_4to3');
-  const prevBtn_5to4 = document.getElementById('prevBtn_5to4');
-
-  nextBtn_1to2.addEventListener("click", () => {pgAtoB(1,2)});
-  nextBtn_2to3.addEventListener("click", () => {pgAtoB(2,3)});
-  nextBtn_3to4.addEventListener("click", () => {pgAtoB(3,4)});
-  nextBtn_4to5.addEventListener("click", () => {pgAtoB(4,5)});
-
-  prevBtn_2to1.addEventListener("click", () => {pgAtoB(2,1)});
-  prevBtn_3to2.addEventListener("click", () => {pgAtoB(3,2)});
-  prevBtn_4to3.addEventListener("click", () => {pgAtoB(4,3)});
-  prevBtn_5to4.addEventListener("click", () => {pgAtoB(5,4)});
-}
-
-let homeToggle = false;
-let machineToggle = false;
-let saveToggle = false;
-let secondbar = document.getElementById('secondbar');
-
-function closeMenu() {
-  if (homeToggle || machineToggle || saveToggle) {
-    secondbar.style.transform = "translate(-240px)";
-    homeToggle = false;
-    machineToggle = false;
-    saveToggle = false;
-  }
-  document.querySelector('.active')?.classList.remove('active');
-}
-
-function clearMenu() {
-  var i;
-  var x = document.getElementsByClassName('dropdown');
-  for (i = 0; i < x.length; i++) {  
-    x[i].hidden = true;
-  }
-  document.querySelector('.active')?.classList.remove('active');
-}
-
-/**
- * Toggle menu open or closed
- * @param {Object} classname - name of class associated with icon
- * @returns {boolean} true if menu was opened, false if closed
- */
-function toggleMenu(classname) {
-  // if closed, open the appropriate menu, or close if clicked on the same icon
-  if (classname === 'home' && !homeToggle) {
-    window.requestAnimationFrame(function(){
-      secondbar.style.transform = "translate(0vw)"; 
-    });
-    document.getElementById('secondbar').hidden = false;
-    var x = document.getElementsByClassName(classname);
-    for (var i = 0; i < x.length; i++) {
-      x[i].hidden = false;
-    }
-    homeToggle = true;
-    machineToggle = false;
-    saveToggle = false;
-    return true;
-  }
-  else if (classname === 'settings' && !machineToggle) {
-    window.requestAnimationFrame(function(){
-      secondbar.style.transform = "translate(0vw)"; 
-    });
-    document.getElementById('secondbar').hidden = false;
-    var i;
-    var x = document.getElementsByClassName(classname);
-    for (i = 0; i < x.length; i++) {
-      x[i].hidden = false;
-    }
-    homeToggle = false;
-    machineToggle = true;
-    saveToggle = false;
-    return true;
-  }
-  else if (classname === 'save' && !saveToggle) {
-    window.requestAnimationFrame(function(){
-      secondbar.style.transform = "translate(0vw)"; 
-    });
-    document.getElementById('secondbar').hidden = false;
-    var i;
-    var x = document.getElementsByClassName(classname);
-    for (i = 0; i < x.length; i++) {
-      x[i].hidden = false;
-    }
-    homeToggle = false;
-    machineToggle = false;
-    saveToggle = true;
-    return true;
-  }
-  else {
-    closeMenu();
-    return false;
-  }
-}
-
-
-/** Global variable that records current page number of tutorial
- *  Used for funtions openPopup() closePopup() and pgAtoB() */
-var currentPg = 1;
-// Function to open the pop-up
-function openPopup() {
-  //set this to the total number of pages to avoid bug
-  //updated to 5 pages tutorial 7/24/2023
-  currentPg = 5;
-  const overlay = document.getElementById('overlay');
-  overlay.style.display = 'block';
-  const popup = document.querySelector('.popup');
-  popup.style.display = 'block';
-  for(let i=currentPg; i>1 ; i--){
-    pgAtoB(i,i-1);
-  }
-}
-
-// Function to close the pop-up
-function closePopup() {
-  pgAtoB(currentPg,1);
-  var popup = document.querySelector('.popup');
-  popup.style.display = 'none';
-  const overlay = document.getElementById('overlay');
-  overlay.style.display = 'none';
-}
-
-//updated on function when clicking on an icon 5/16/2023
-function expandIcon(nameOfClass){
-  clearMenu();
-  var headerName = 'none';
-  var currIcon;
-  switch(nameOfClass){
-    case 'home':
-      headerName = 'Home';
-      currIcon = document.getElementById("homeIcon");
-      break;
-    case 'settings':
-      headerName = 'Machines';
-      currIcon = document.getElementById('machineIcon');
-      break;
-    case 'save':
-      headerName = 'Save';
-      currIcon = document.getElementById('saveIcon');
-      break;
-    case 'bug':
-      headerName = 'Bug';
-      currIcon = document.getElementById('bugIcon');
-      break;
-    }
-   if (toggleMenu(nameOfClass)) {
-    const header = document.querySelector('#secondBarHeaderTitle > h1');
-    header.textContent = headerName;
-    currIcon.classList.add('active');
-  }
-  else {
-    document.querySelector('.active')?.classList.remove('active');
-  }
-}
-
-// const navLinks = document.querySelectorAll('.sidenavBtn');
-// navLinks.forEach(act => {
-//   act.addEventListener('click', () => {
-//     document.querySelector('.active')?.classList.remove('active');
-//     this.classList.add('active');
-//   })
-// });
-function pgAtoB(a,b){
-  var pgName = 'pg';
-  var nameA = pgName + a;
-  var nameB = pgName + b;
-  const pgA = document.getElementById(nameA);
-  const pgB = document.getElementById(nameB);
-  pgA.style.display = "none";
-  pgB.style.display = "block";
-  currentPg = b;
-}
-
-function redirectToBugReport() {
-  window.open('https://github.com/flapjs/vanilla/issues', '_blank');
-  //below is code for opening bug report in current tab of browser
-  //window.location.href = 'https://github.com/flapjs/vanilla/issues';
 }
