+++ scripts/util.js
@@ -1,5 +1,7 @@
 /** @module util */
 
+import * as consts from './consts.js';
+
 export class Queue {
   constructor() {
     this.arr = [];
@@ -48,4 +50,23 @@ export function deep_equal(obj1, obj2) {
     }
   }
   return true;
+}
+
+/**
+ * helper function to abstract away double pressing a key
+ * @param {string} key - ex. KeyZ, KeyA
+ * @param {Function} callback - a function to be called when double click happens
+ */
+export function on_double_press(key, callback) {
+ let last_time = 0;
+ document.addEventListener('keypress', e => {
+   if (e.code === key) {
+     if (e.timeStamp-last_time < consts.DOUBLE_PRESS_TIME) {
+       callback();
+       last_time = 0;  // prevent triple click
+     } else {
+       last_time = e.timeStamp;
+     }
+   }
+ });
 }
