+++ styles/index.css
@@ -21,8 +21,16 @@ canvas {
   background-color:#f2f0f0;
 }
 
+button {
+  border-radius: 5px;
+}
+
 button:hover {
   cursor: pointer;
+}
+
+input {
+  border-radius: 5px;
 }
 
 /***** Graph Context Menu *****/
@@ -42,7 +50,6 @@ button:hover {
 
 .context_menu button:hover {
   background-color: #c7c2c2;
-  border-radius: 5px;
   transition: all .15s ease-in-out;
 }
 
@@ -122,7 +129,6 @@ button:hover {
   font-size: 13px;
   color:gold;
   background-color: #3a468f;
-  border-radius: 5px;
   transition: all .5s;
 }
 
@@ -151,10 +157,10 @@ button:hover {
 select {
   position: relative;
   font-size: 10px;
-  left: 1.5vw;
+  left: 1.4vw;
   padding: 10px 0px 10px 10px;
   border: 2px solid #ccc;
-  border-radius: 10px;
+  border-radius: 5px;
 }
 
 /* Darken background window when popup is open */
